Added dependencies files

+++ store/models.py
@@ -11,9 +11,10 @@ class Customer(models.Model):
 	def __str__(self):
 		return self.name
 
+
 class Product(models.Model):
 	name = models.CharField(max_length=200)
-	price = models.DecimalField(max_digits=7,decimal_places=2)
+	price = models.FloatField()
 	digital = models.BooleanField(default=False,null=True, blank=True)
 	image = models.ImageField(null=True, blank=True)
